stamp-api: Optimise rendering and fix async issue
Previously, we cached HTMLImageElement instances for each sticker to
reuse when redrawing. The stickers are SVGs which meant that drawImage
had to rasterize it every time. In this PR, I changed the caching to
store instances of HTMLCanvasElement with the images rendered. That
seems to improve the performance quite a bit.

While looking at it, I discovered an issue with the way we retrieved images
'synchronously' on-demand when we needed them to render. The while loop
we used was locking the event loop for a specified timeout preventing
images from being loaded rather than waiting for them to be loaded.

To work around this issue, I made the userCode run asynchronously using
the async/await keywords. This seems to work without us having to
restructure the VM in a substantial way and gives us the option to do
other async things in userCode in the future.